+++ lib/articleUrl.ts
@@ -1,0 +1,5 @@
+export function articleUrl(publicSiteUrl: string, contentPath: string): string {
+  const origin = new URL(publicSiteUrl).origin
+  const path = contentPath.split(/[?#]/, 1)[0].replace(/^\/+/, '')
+  return `${origin}/${path}`
+}